refactoring native/special function definitions out of lib.rs
