+++ blockstack-vm/src/lib.rs
@@ -1,18 +1,13 @@
 pub mod types;
 pub mod representations;
 
+mod functions;
+
 use std::collections::HashMap;
 use types::ValueType;
 use types::CallableType;
 use types::DefinedFunction;
 use representations::SymbolicExpression;
-
-fn type_force_integer(value: &ValueType) -> u64 {
-    match *value {
-        ValueType::IntType(int) => int,
-        _ => panic!("Not an integer")
-    }
-}
 
 pub struct Context <'a> {
     pub parent: Option< &'a Context<'a>>,
@@ -54,47 +49,6 @@ impl <'a> Context <'a> {
     }
 }
 
-fn native_add(args: &[ValueType]) -> ValueType {
-    let parsed_args = args.iter().map(|x| type_force_integer(x));
-    let result = parsed_args.fold(0, |acc, x| acc + x);
-    ValueType::IntType(result)
-}
-
-fn native_eq(args: &[ValueType]) -> ValueType {
-    // TODO: this currently uses the derived equality checks of ValueType,
-    //   however, that's probably not how we want to implement equality
-    //   checks on the ::ListTypes
-    if args.len() < 2 {
-        ValueType::BoolType(true)
-    } else {
-        let first = &args[0];
-        let result = args.iter().fold(true, |acc, x| acc && (*x == *first));
-        ValueType::BoolType(result)
-    }
-}
-
-fn special_if(args: &[SymbolicExpression], context: &Context) -> ValueType {
-    if !(args.len() == 2 || args.len() == 3) {
-        panic!("Wrong number of arguments to if");
-    }
-    // handle the conditional clause.
-    let conditional = eval(&args[0], context);
-    match conditional {
-        ValueType::BoolType(result) => {
-            if result {
-                eval(&args[1], context)
-            } else {
-                if args.len() == 3 {
-                    eval(&args[2], context)
-                } else {
-                    ValueType::VoidType
-                }
-            }
-        },
-        _ => panic!("Conditional argument must evaluate to BoolType")
-    }
-}
-
 fn lookup_variable(name: &str, context: &Context) -> ValueType {
     // first off, are we talking about a constant?
     if name.starts_with(char::is_numeric) {
@@ -111,10 +65,8 @@ fn lookup_variable(name: &str, context: &Context) -> ValueType {
 }
 
 fn lookup_function<'a> (name: &str, context: &'a Context)-> CallableType<'a> {
-    match name {
-        "+" => CallableType::NativeFunction(&native_add),
-        "eq?" => CallableType::NativeFunction(&native_eq),
-        "if" => CallableType::SpecialFunction(&special_if),
+    match functions::lookup_reserved_functions(name) {
+        Some(result) => result,
         _ => {
             match context.lookup_function(name) {
                 Some(func) => {
